refactor: use the official `kube_godegen.sh` script as recommended by repo
