+++ hack/tools.go
@@ -1,0 +1,8 @@
+//go:build tools
+
+package hack
+
+import (
+	_ "k8s.io/code-generator/cmd/client-gen"
+)
+
