+++ internal/controller/variableset/variableset_controller.go
@@ -19,7 +19,7 @@ package variableset
 import (
 	"context"
 
-	magosprojectiov1alpha1 "github.com/magosproject/magos/types/v1alpha1"
+	magosprojectiov1alpha1 "github.com/magosproject/magos/types/magosproject/v1alpha1"
 	"k8s.io/apimachinery/pkg/runtime"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
