+++ internal/controller/workspace/workspace_controller.go
@@ -23,7 +23,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/magosproject/magos/types/v1alpha1"
+	"github.com/magosproject/magos/types/magosproject/v1alpha1"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
