+++ types/magosproject/v1alpha1/groupversion_info.go
