+++ types/magosproject/v1alpha1/rollout_types.go
